gnome-keyring: start by using systemd user service
Currently both the systemd and the non-systemd way to bring up the daemon are mixed
in a way that is apparently working but leads to have the keyring enabled with pam
and later on a second instance gets started by systemd socket activation.

Make a decision here and go the systemd way only to avoid this.