move the xwayland patches
gnome-shell does not yet use fractional scaling, so the problem with hidpi
displays and xwayland applications is not as acute as when using current
wlroots. Also, it is quite cumbersome to have to set the xproperties with
xprop on the root window to get the proper scaling.

Therefore avoid patching xwayland for gnome.